refactor(feature/#35): distanceSetting 폴더 만들기
